Changed type of compute invocation and group counting to uint32

+++ VitraeEngine/include/Vitrae/Util/GpuCompute.hpp
@@ -9,12 +9,12 @@ inline constexpr std::size_t GROUP_SIZE_AUTO = 0;
 
 struct GpuComputeSetupParams
 {
-    PropertyGetter<std::size_t> invocationCountX;
-    PropertyGetter<std::size_t> invocationCountY = 1;
-    PropertyGetter<std::size_t> invocationCountZ = 1;
-    PropertyGetter<std::size_t> groupSizeX = GROUP_SIZE_AUTO;
-    PropertyGetter<std::size_t> groupSizeY = GROUP_SIZE_AUTO;
-    PropertyGetter<std::size_t> groupSizeZ = GROUP_SIZE_AUTO;
+    PropertyGetter<std::uint32_t> invocationCountX;
+    PropertyGetter<std::uint32_t> invocationCountY = 1;
+    PropertyGetter<std::uint32_t> invocationCountZ = 1;
+    PropertyGetter<std::uint32_t> groupSizeX = GROUP_SIZE_AUTO;
+    PropertyGetter<std::uint32_t> groupSizeY = GROUP_SIZE_AUTO;
+    PropertyGetter<std::uint32_t> groupSizeZ = GROUP_SIZE_AUTO;
     bool allowOutOfBoundsCompute = false;
 };
